Update README.md
Included a link to the Discord channel where readers will find it as they read through where and how to contribute to the project.

This update is consistent with a norm that I have seen elsewhere. The link that is already in place, does not come up when searching the page for "Discord" and might otherwise be difficult to find (as was my experience).